Move tests tests from frame/response/result.rs to deserialize/value_tests
Those tests only test CqlValue deserialization, and we want to have
serialization / deserialization tests grouped together.
