+++ scylla-cql/src/deserialize/value_tests.rs
@@ -7,10 +7,12 @@ use std::borrow::Cow;
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::fmt::Debug;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
+use std::str::FromStr;
 use std::sync::Arc;
 
 use crate::deserialize::value::{TupleDeserializationErrorKind, TupleTypeCheckErrorKind};
 use crate::deserialize::{DeserializationError, FrameSlice, TypeCheckError};
+use crate::frame::response::result::NativeType::*;
 use crate::frame::response::result::{
     CollectionType, ColumnType, CqlValue, NativeType, UserDefinedType,
 };
@@ -2364,4 +2366,890 @@ fn metadata_does_not_bound_deserialized_values() {
             decoded_udt_res,
         )
     };
+}
+
+// Tests migrated from old frame/value_tests.rs file
+
+#[test]
+fn test_deserialize_text_types() {
+    let buf: Vec<u8> = vec![0x41];
+    let int_slice = &mut &buf[..];
+    let ascii_serialized = super::deser_cql_value(&ColumnType::Native(Ascii), int_slice).unwrap();
+    let text_serialized = super::deser_cql_value(&ColumnType::Native(Text), int_slice).unwrap();
+    assert_eq!(ascii_serialized, CqlValue::Ascii("A".to_string()));
+    assert_eq!(text_serialized, CqlValue::Text("A".to_string()));
+}
+
+#[test]
+fn test_deserialize_uuid_inet_types() {
+    let my_uuid = Uuid::parse_str("00000000000000000000000000000001").unwrap();
+
+    let uuid_buf: Vec<u8> = vec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1];
+    let uuid_slice = &mut &uuid_buf[..];
+    let uuid_serialize = super::deser_cql_value(&ColumnType::Native(Uuid), uuid_slice).unwrap();
+    assert_eq!(uuid_serialize, CqlValue::Uuid(my_uuid));
+
+    let my_timeuuid = CqlTimeuuid::from_str("00000000000000000000000000000001").unwrap();
+    let time_uuid_serialize =
+        super::deser_cql_value(&ColumnType::Native(Timeuuid), uuid_slice).unwrap();
+    assert_eq!(time_uuid_serialize, CqlValue::Timeuuid(my_timeuuid));
+
+    let my_ip = "::1".parse().unwrap();
+    let ip_buf: Vec<u8> = vec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1];
+    let ip_slice = &mut &ip_buf[..];
+    let ip_serialize = super::deser_cql_value(&ColumnType::Native(Inet), ip_slice).unwrap();
+    assert_eq!(ip_serialize, CqlValue::Inet(my_ip));
+
+    let max_ip = "ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff".parse().unwrap();
+    let max_ip_buf: Vec<u8> = vec![
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
+    ];
+    let max_ip_slice = &mut &max_ip_buf[..];
+    let max_ip_serialize = super::deser_cql_value(&ColumnType::Native(Inet), max_ip_slice).unwrap();
+    assert_eq!(max_ip_serialize, CqlValue::Inet(max_ip));
+}
+
+#[test]
+fn test_floating_points() {
+    let float: f32 = 0.5;
+    let double: f64 = 2.0;
+
+    let float_buf: Vec<u8> = vec![63, 0, 0, 0];
+    let float_slice = &mut &float_buf[..];
+    let float_serialize = super::deser_cql_value(&ColumnType::Native(Float), float_slice).unwrap();
+    assert_eq!(float_serialize, CqlValue::Float(float));
+
+    let double_buf: Vec<u8> = vec![64, 0, 0, 0, 0, 0, 0, 0];
+    let double_slice = &mut &double_buf[..];
+    let double_serialize =
+        super::deser_cql_value(&ColumnType::Native(Double), double_slice).unwrap();
+    assert_eq!(double_serialize, CqlValue::Double(double));
+}
+
+#[cfg(any(feature = "num-bigint-03", feature = "num-bigint-04"))]
+struct VarintTestCase {
+    value: i32,
+    encoding: Vec<u8>,
+}
+
+#[cfg(any(feature = "num-bigint-03", feature = "num-bigint-04"))]
+fn varint_test_cases_from_spec() -> Vec<VarintTestCase> {
+    /*
+        Table taken from CQL Binary Protocol v4 spec
+
+        Value | Encoding
+        ------|---------
+            0 |     0x00
+            1 |     0x01
+          127 |     0x7F
+          128 |   0x0080
+          129 |   0x0081
+           -1 |     0xFF
+         -128 |     0x80
+         -129 |   0xFF7F
+    */
+    vec![
+        VarintTestCase {
+            value: 0,
+            encoding: vec![0x00],
+        },
+        VarintTestCase {
+            value: 1,
+            encoding: vec![0x01],
+        },
+        VarintTestCase {
+            value: 127,
+            encoding: vec![0x7F],
+        },
+        VarintTestCase {
+            value: 128,
+            encoding: vec![0x00, 0x80],
+        },
+        VarintTestCase {
+            value: 129,
+            encoding: vec![0x00, 0x81],
+        },
+        VarintTestCase {
+            value: -1,
+            encoding: vec![0xFF],
+        },
+        VarintTestCase {
+            value: -128,
+            encoding: vec![0x80],
+        },
+        VarintTestCase {
+            value: -129,
+            encoding: vec![0xFF, 0x7F],
+        },
+    ]
+}
+
+#[cfg(feature = "num-bigint-03")]
+#[test]
+fn test_bigint03() {
+    use num_bigint_03::ToBigInt;
+
+    let tests = varint_test_cases_from_spec();
+
+    for t in tests.iter() {
+        let value = super::deser_cql_value(&ColumnType::Native(Varint), &mut &*t.encoding).unwrap();
+        assert_eq!(CqlValue::Varint(t.value.to_bigint().unwrap().into()), value);
+    }
+}
+
+#[cfg(feature = "num-bigint-04")]
+#[test]
+fn test_bigint04() {
+    use num_bigint_04::ToBigInt;
+
+    let tests = varint_test_cases_from_spec();
+
+    for t in tests.iter() {
+        let value = super::deser_cql_value(&ColumnType::Native(Varint), &mut &*t.encoding).unwrap();
+        assert_eq!(CqlValue::Varint(t.value.to_bigint().unwrap().into()), value);
+    }
+}
+
+#[cfg(feature = "bigdecimal-04")]
+#[test]
+fn test_decimal() {
+    use bigdecimal_04::BigDecimal;
+    struct Test<'a> {
+        value: BigDecimal,
+        encoding: &'a [u8],
+    }
+
+    let tests = [
+        Test {
+            value: BigDecimal::from_str("-1.28").unwrap(),
+            encoding: &[0x0, 0x0, 0x0, 0x2, 0x80],
+        },
+        Test {
+            value: BigDecimal::from_str("1.29").unwrap(),
+            encoding: &[0x0, 0x0, 0x0, 0x2, 0x0, 0x81],
+        },
+        Test {
+            value: BigDecimal::from_str("0").unwrap(),
+            encoding: &[0x0, 0x0, 0x0, 0x0, 0x0],
+        },
+        Test {
+            value: BigDecimal::from_str("123").unwrap(),
+            encoding: &[0x0, 0x0, 0x0, 0x0, 0x7b],
+        },
+    ];
+
+    for t in tests.iter() {
+        let value =
+            super::deser_cql_value(&ColumnType::Native(Decimal), &mut &*t.encoding).unwrap();
+        assert_eq!(
+            CqlValue::Decimal(t.value.clone().try_into().unwrap()),
+            value
+        );
+    }
+}
+
+#[test]
+fn test_deserialize_counter() {
+    let counter: Vec<u8> = vec![0, 0, 0, 0, 0, 0, 1, 0];
+    let counter_slice = &mut &counter[..];
+    let counter_serialize =
+        super::deser_cql_value(&ColumnType::Native(NativeType::Counter), counter_slice).unwrap();
+    assert_eq!(counter_serialize, CqlValue::Counter(Counter(256)));
+}
+
+#[test]
+fn test_deserialize_blob() {
+    let blob: Vec<u8> = vec![0, 1, 2, 3];
+    let blob_slice = &mut &blob[..];
+    let blob_serialize = super::deser_cql_value(&ColumnType::Native(Blob), blob_slice).unwrap();
+    assert_eq!(blob_serialize, CqlValue::Blob(blob));
+}
+
+#[test]
+fn test_deserialize_bool() {
+    let bool_buf: Vec<u8> = vec![0x00];
+    let bool_slice = &mut &bool_buf[..];
+    let bool_serialize = super::deser_cql_value(&ColumnType::Native(Boolean), bool_slice).unwrap();
+    assert_eq!(bool_serialize, CqlValue::Boolean(false));
+
+    let bool_buf: Vec<u8> = vec![0x01];
+    let bool_slice = &mut &bool_buf[..];
+    let bool_serialize = super::deser_cql_value(&ColumnType::Native(Boolean), bool_slice).unwrap();
+    assert_eq!(bool_serialize, CqlValue::Boolean(true));
+}
+
+#[test]
+fn test_deserialize_int_types() {
+    let int_buf: Vec<u8> = vec![0, 0, 0, 4];
+    let int_slice = &mut &int_buf[..];
+    let int_serialized = super::deser_cql_value(&ColumnType::Native(Int), int_slice).unwrap();
+    assert_eq!(int_serialized, CqlValue::Int(4));
+
+    let smallint_buf: Vec<u8> = vec![0, 4];
+    let smallint_slice = &mut &smallint_buf[..];
+    let smallint_serialized =
+        super::deser_cql_value(&ColumnType::Native(SmallInt), smallint_slice).unwrap();
+    assert_eq!(smallint_serialized, CqlValue::SmallInt(4));
+
+    let tinyint_buf: Vec<u8> = vec![4];
+    let tinyint_slice = &mut &tinyint_buf[..];
+    let tinyint_serialized =
+        super::deser_cql_value(&ColumnType::Native(TinyInt), tinyint_slice).unwrap();
+    assert_eq!(tinyint_serialized, CqlValue::TinyInt(4));
+
+    let bigint_buf: Vec<u8> = vec![0, 0, 0, 0, 0, 0, 0, 4];
+    let bigint_slice = &mut &bigint_buf[..];
+    let bigint_serialized =
+        super::deser_cql_value(&ColumnType::Native(BigInt), bigint_slice).unwrap();
+    assert_eq!(bigint_serialized, CqlValue::BigInt(4));
+}
+
+#[test]
+fn test_list_from_cql() {
+    let my_vec: Vec<CqlValue> = vec![CqlValue::Int(20), CqlValue::Int(2), CqlValue::Int(13)];
+
+    let cql: CqlValue = CqlValue::List(my_vec);
+    let decoded = cql.into_vec().unwrap();
+
+    assert_eq!(decoded[0], CqlValue::Int(20));
+    assert_eq!(decoded[1], CqlValue::Int(2));
+    assert_eq!(decoded[2], CqlValue::Int(13));
+}
+
+#[test]
+fn test_set_from_cql() {
+    let my_vec: Vec<CqlValue> = vec![CqlValue::Int(20), CqlValue::Int(2), CqlValue::Int(13)];
+
+    let cql: CqlValue = CqlValue::Set(my_vec);
+    let decoded = cql.as_set().unwrap();
+
+    assert_eq!(decoded[0], CqlValue::Int(20));
+    assert_eq!(decoded[1], CqlValue::Int(2));
+    assert_eq!(decoded[2], CqlValue::Int(13));
+}
+
+#[test]
+fn test_map_from_cql() {
+    let my_vec: Vec<(CqlValue, CqlValue)> = vec![
+        (CqlValue::Int(20), CqlValue::Int(21)),
+        (CqlValue::Int(2), CqlValue::Int(3)),
+    ];
+
+    let cql: CqlValue = CqlValue::Map(my_vec);
+
+    // Test borrowing.
+    let decoded = cql.as_map().unwrap();
+
+    assert_eq!(CqlValue::Int(20), decoded[0].0);
+    assert_eq!(CqlValue::Int(21), decoded[0].1);
+
+    assert_eq!(CqlValue::Int(2), decoded[1].0);
+    assert_eq!(CqlValue::Int(3), decoded[1].1);
+
+    // Test taking the ownership.
+    let decoded = cql.into_pair_vec().unwrap();
+
+    assert_eq!(CqlValue::Int(20), decoded[0].0);
+    assert_eq!(CqlValue::Int(21), decoded[0].1);
+
+    assert_eq!(CqlValue::Int(2), decoded[1].0);
+    assert_eq!(CqlValue::Int(3), decoded[1].1);
+}
+
+#[test]
+fn test_udt_from_cql() {
+    let my_fields: Vec<(String, Option<CqlValue>)> = vec![
+        ("fst".to_string(), Some(CqlValue::Int(10))),
+        ("snd".to_string(), Some(CqlValue::Boolean(true))),
+    ];
+
+    let cql: CqlValue = CqlValue::UserDefinedType {
+        keyspace: "".to_string(),
+        name: "".to_string(),
+        fields: my_fields,
+    };
+
+    // Test borrowing.
+    let decoded = cql.as_udt().unwrap();
+
+    assert_eq!("fst".to_string(), decoded[0].0);
+    assert_eq!(Some(CqlValue::Int(10)), decoded[0].1);
+
+    assert_eq!("snd".to_string(), decoded[1].0);
+    assert_eq!(Some(CqlValue::Boolean(true)), decoded[1].1);
+
+    let decoded = cql.into_udt_pair_vec().unwrap();
+
+    assert_eq!("fst".to_string(), decoded[0].0);
+    assert_eq!(Some(CqlValue::Int(10)), decoded[0].1);
+
+    assert_eq!("snd".to_string(), decoded[1].0);
+    assert_eq!(Some(CqlValue::Boolean(true)), decoded[1].1);
+}
+
+#[test]
+fn test_deserialize_date() {
+    // Date is correctly parsed from a 4 byte array
+    let four_bytes: [u8; 4] = [12, 23, 34, 45];
+    let date: CqlValue =
+        super::deser_cql_value(&ColumnType::Native(Date), &mut four_bytes.as_ref()).unwrap();
+    assert_eq!(
+        date,
+        CqlValue::Date(CqlDate(u32::from_be_bytes(four_bytes)))
+    );
+
+    // Date is parsed as u32 not i32, u32::MAX is u32::MAX
+    let date: CqlValue = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut u32::MAX.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+    assert_eq!(date, CqlValue::Date(CqlDate(u32::MAX)));
+
+    // Trying to parse a 0, 3 or 5 byte array fails
+    super::deser_cql_value(&ColumnType::Native(Date), &mut [].as_ref()).unwrap();
+    super::deser_cql_value(&ColumnType::Native(Date), &mut [1, 2, 3].as_ref()).unwrap_err();
+    super::deser_cql_value(&ColumnType::Native(Date), &mut [1, 2, 3, 4, 5].as_ref()).unwrap_err();
+
+    // Deserialize unix epoch
+    let unix_epoch_bytes = 2_u32.pow(31).to_be_bytes();
+
+    let date =
+        super::deser_cql_value(&ColumnType::Native(Date), &mut unix_epoch_bytes.as_ref()).unwrap();
+    assert_eq!(date.as_cql_date(), Some(CqlDate(1 << 31)));
+
+    // 2^31 - 30 when converted to NaiveDate is 1969-12-02
+    let before_epoch = CqlDate((1 << 31) - 30);
+    let date: CqlValue = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut ((1_u32 << 31) - 30).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_cql_date(), Some(before_epoch));
+
+    // 2^31 + 30 when converted to NaiveDate is 1970-01-31
+    let after_epoch = CqlDate((1 << 31) + 30);
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut ((1_u32 << 31) + 30).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_cql_date(), Some(after_epoch));
+
+    // Min date
+    let min_date = CqlDate(u32::MIN);
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut u32::MIN.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+    assert_eq!(date.as_cql_date(), Some(min_date));
+
+    // Max date
+    let max_date = CqlDate(u32::MAX);
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut u32::MAX.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+    assert_eq!(date.as_cql_date(), Some(max_date));
+}
+
+#[cfg(feature = "chrono-04")]
+#[test]
+fn test_naive_date_04_from_cql() {
+    use chrono_04::NaiveDate;
+
+    // 2^31 when converted to NaiveDate is 1970-01-01
+    let unix_epoch = NaiveDate::from_ymd_opt(1970, 1, 1).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut (1u32 << 31).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_naive_date_04(), Some(unix_epoch));
+
+    // 2^31 - 30 when converted to NaiveDate is 1969-12-02
+    let before_epoch = NaiveDate::from_ymd_opt(1969, 12, 2).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut ((1u32 << 31) - 30).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_naive_date_04(), Some(before_epoch));
+
+    // 2^31 + 30 when converted to NaiveDate is 1970-01-31
+    let after_epoch = NaiveDate::from_ymd_opt(1970, 1, 31).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut ((1u32 << 31) + 30).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_naive_date_04(), Some(after_epoch));
+
+    // 0 and u32::MAX are out of NaiveDate range, fails with an error, not panics
+    assert_eq!(
+        super::deser_cql_value(&ColumnType::Native(Date), &mut 0_u32.to_be_bytes().as_ref())
+            .unwrap()
+            .as_naive_date_04(),
+        None
+    );
+
+    assert_eq!(
+        super::deser_cql_value(
+            &ColumnType::Native(Date),
+            &mut u32::MAX.to_be_bytes().as_ref()
+        )
+        .unwrap()
+        .as_naive_date_04(),
+        None
+    );
+}
+
+#[cfg(feature = "time-03")]
+#[test]
+fn test_date_03_from_cql() {
+    use time_03::Date;
+    use time_03::Month::*;
+
+    // 2^31 when converted to time_03::Date is 1970-01-01
+    let unix_epoch = Date::from_calendar_date(1970, January, 1).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut (1u32 << 31).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_date_03(), Some(unix_epoch));
+
+    // 2^31 - 30 when converted to time_03::Date is 1969-12-02
+    let before_epoch = Date::from_calendar_date(1969, December, 2).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut ((1u32 << 31) - 30).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_date_03(), Some(before_epoch));
+
+    // 2^31 + 30 when converted to time_03::Date is 1970-01-31
+    let after_epoch = Date::from_calendar_date(1970, January, 31).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Date),
+        &mut ((1u32 << 31) + 30).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_date_03(), Some(after_epoch));
+
+    // 0 and u32::MAX are out of NaiveDate range, fails with an error, not panics
+    assert_eq!(
+        super::deser_cql_value(&ColumnType::Native(Date), &mut 0_u32.to_be_bytes().as_ref())
+            .unwrap()
+            .as_date_03(),
+        None
+    );
+
+    assert_eq!(
+        super::deser_cql_value(
+            &ColumnType::Native(Date),
+            &mut u32::MAX.to_be_bytes().as_ref()
+        )
+        .unwrap()
+        .as_date_03(),
+        None
+    );
+}
+
+#[test]
+fn test_deserialize_time() {
+    // Time is an i64 - nanoseconds since midnight
+    // in range 0..=86399999999999
+
+    let max_time: i64 = 24 * 60 * 60 * 1_000_000_000 - 1;
+    assert_eq!(max_time, 86399999999999);
+
+    // Check that basic values are deserialized correctly
+    for test_val in [0, 1, 18463, max_time].iter() {
+        let bytes: [u8; 8] = test_val.to_be_bytes();
+        let cql_value: CqlValue =
+            super::deser_cql_value(&ColumnType::Native(Time), &mut &bytes[..]).unwrap();
+        assert_eq!(cql_value, CqlValue::Time(CqlTime(*test_val)));
+    }
+
+    // Negative values cause an error
+    // Values bigger than 86399999999999 cause an error
+    for test_val in [-1, i64::MIN, max_time + 1, i64::MAX].iter() {
+        let bytes: [u8; 8] = test_val.to_be_bytes();
+        super::deser_cql_value(&ColumnType::Native(Time), &mut &bytes[..]).unwrap_err();
+    }
+}
+
+#[cfg(feature = "chrono-04")]
+#[test]
+fn test_naive_time_04_from_cql() {
+    use chrono_04::NaiveTime;
+
+    // 0 when converted to NaiveTime is 0:0:0.0
+    let midnight = NaiveTime::from_hms_nano_opt(0, 0, 0, 0).unwrap();
+    let time = super::deser_cql_value(
+        &ColumnType::Native(Time),
+        &mut (0i64).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(time.as_naive_time_04(), Some(midnight));
+
+    // 10:10:30.500,000,001
+    let (h, m, s, n) = (10, 10, 30, 500_000_001);
+    let midnight = NaiveTime::from_hms_nano_opt(h, m, s, n).unwrap();
+    let time = super::deser_cql_value(
+        &ColumnType::Native(Time),
+        &mut ((h as i64 * 3600 + m as i64 * 60 + s as i64) * 1_000_000_000 + n as i64)
+            .to_be_bytes()
+            .as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(time.as_naive_time_04(), Some(midnight));
+
+    // 23:59:59.999,999,999
+    let (h, m, s, n) = (23, 59, 59, 999_999_999);
+    let midnight = NaiveTime::from_hms_nano_opt(h, m, s, n).unwrap();
+    let time = super::deser_cql_value(
+        &ColumnType::Native(Time),
+        &mut ((h as i64 * 3600 + m as i64 * 60 + s as i64) * 1_000_000_000 + n as i64)
+            .to_be_bytes()
+            .as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(time.as_naive_time_04(), Some(midnight));
+}
+
+#[cfg(feature = "time-03")]
+#[test]
+fn test_primitive_time_03_from_cql() {
+    use time_03::Time;
+
+    // 0 when converted to NaiveTime is 0:0:0.0
+    let midnight = Time::from_hms_nano(0, 0, 0, 0).unwrap();
+    let time = super::deser_cql_value(
+        &ColumnType::Native(Time),
+        &mut (0i64).to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(time.as_time_03(), Some(midnight));
+
+    // 10:10:30.500,000,001
+    let (h, m, s, n) = (10, 10, 30, 500_000_001);
+    let midnight = Time::from_hms_nano(h, m, s, n).unwrap();
+    let time = super::deser_cql_value(
+        &ColumnType::Native(Time),
+        &mut ((h as i64 * 3600 + m as i64 * 60 + s as i64) * 1_000_000_000 + n as i64)
+            .to_be_bytes()
+            .as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(time.as_time_03(), Some(midnight));
+
+    // 23:59:59.999,999,999
+    let (h, m, s, n) = (23, 59, 59, 999_999_999);
+    let midnight = Time::from_hms_nano(h, m, s, n).unwrap();
+    let time = super::deser_cql_value(
+        &ColumnType::Native(Time),
+        &mut ((h as i64 * 3600 + m as i64 * 60 + s as i64) * 1_000_000_000 + n as i64)
+            .to_be_bytes()
+            .as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(time.as_time_03(), Some(midnight));
+}
+
+#[test]
+fn test_timestamp_deserialize() {
+    // Timestamp is an i64 - milliseconds since unix epoch
+
+    // Check that test values are deserialized correctly
+    for test_val in &[0, -1, 1, 74568745, -4584658, i64::MIN, i64::MAX] {
+        let bytes: [u8; 8] = test_val.to_be_bytes();
+        let cql_value: CqlValue =
+            super::deser_cql_value(&ColumnType::Native(Timestamp), &mut &bytes[..]).unwrap();
+        assert_eq!(cql_value, CqlValue::Timestamp(CqlTimestamp(*test_val)));
+    }
+}
+
+#[cfg(feature = "chrono-04")]
+#[test]
+fn test_datetime_04_from_cql() {
+    use chrono_04::{DateTime, NaiveDate, NaiveDateTime, NaiveTime};
+
+    // 0 when converted to DateTime is 1970-01-01 0:00:00.00
+    let unix_epoch = DateTime::from_timestamp(0, 0).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Timestamp),
+        &mut 0i64.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_datetime_04(), Some(unix_epoch));
+
+    // When converted to NaiveDateTime, this is 1969-12-01 11:29:29.5
+    let timestamp: i64 = -((((30 * 24 + 12) * 60 + 30) * 60 + 30) * 1000 + 500);
+    let before_epoch = NaiveDateTime::new(
+        NaiveDate::from_ymd_opt(1969, 12, 1).unwrap(),
+        NaiveTime::from_hms_milli_opt(11, 29, 29, 500).unwrap(),
+    )
+    .and_utc();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Timestamp),
+        &mut timestamp.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_datetime_04(), Some(before_epoch));
+
+    // when converted to NaiveDateTime, this is is 1970-01-31 12:30:30.5
+    let timestamp: i64 = (((30 * 24 + 12) * 60 + 30) * 60 + 30) * 1000 + 500;
+    let after_epoch = NaiveDateTime::new(
+        NaiveDate::from_ymd_opt(1970, 1, 31).unwrap(),
+        NaiveTime::from_hms_milli_opt(12, 30, 30, 500).unwrap(),
+    )
+    .and_utc();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Timestamp),
+        &mut timestamp.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_datetime_04(), Some(after_epoch));
+
+    // 0 and u32::MAX are out of NaiveDate range, fails with an error, not panics
+    assert_eq!(
+        super::deser_cql_value(
+            &ColumnType::Native(Timestamp),
+            &mut i64::MIN.to_be_bytes().as_ref()
+        )
+        .unwrap()
+        .as_datetime_04(),
+        None
+    );
+
+    assert_eq!(
+        super::deser_cql_value(
+            &ColumnType::Native(Timestamp),
+            &mut i64::MAX.to_be_bytes().as_ref()
+        )
+        .unwrap()
+        .as_datetime_04(),
+        None
+    );
+}
+
+#[cfg(feature = "time-03")]
+#[test]
+fn test_offset_datetime_03_from_cql() {
+    use time_03::{Date, Month::*, OffsetDateTime, PrimitiveDateTime, Time};
+
+    // 0 when converted to OffsetDateTime is 1970-01-01 0:00:00.00
+    let unix_epoch = OffsetDateTime::from_unix_timestamp(0).unwrap();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Timestamp),
+        &mut 0i64.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_offset_date_time_03(), Some(unix_epoch));
+
+    // When converted to NaiveDateTime, this is 1969-12-01 11:29:29.5
+    let timestamp: i64 = -((((30 * 24 + 12) * 60 + 30) * 60 + 30) * 1000 + 500);
+    let before_epoch = PrimitiveDateTime::new(
+        Date::from_calendar_date(1969, December, 1).unwrap(),
+        Time::from_hms_milli(11, 29, 29, 500).unwrap(),
+    )
+    .assume_utc();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Timestamp),
+        &mut timestamp.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_offset_date_time_03(), Some(before_epoch));
+
+    // when converted to NaiveDateTime, this is is 1970-01-31 12:30:30.5
+    let timestamp: i64 = (((30 * 24 + 12) * 60 + 30) * 60 + 30) * 1000 + 500;
+    let after_epoch = PrimitiveDateTime::new(
+        Date::from_calendar_date(1970, January, 31).unwrap(),
+        Time::from_hms_milli(12, 30, 30, 500).unwrap(),
+    )
+    .assume_utc();
+    let date = super::deser_cql_value(
+        &ColumnType::Native(Timestamp),
+        &mut timestamp.to_be_bytes().as_ref(),
+    )
+    .unwrap();
+
+    assert_eq!(date.as_offset_date_time_03(), Some(after_epoch));
+
+    // 0 and u32::MAX are out of NaiveDate range, fails with an error, not panics
+    assert_eq!(
+        super::deser_cql_value(
+            &ColumnType::Native(Timestamp),
+            &mut i64::MIN.to_be_bytes().as_ref()
+        )
+        .unwrap()
+        .as_offset_date_time_03(),
+        None
+    );
+
+    assert_eq!(
+        super::deser_cql_value(
+            &ColumnType::Native(Timestamp),
+            &mut i64::MAX.to_be_bytes().as_ref()
+        )
+        .unwrap()
+        .as_offset_date_time_03(),
+        None
+    );
+}
+
+#[test]
+fn test_serialize_empty() {
+    use crate::serialize::value::SerializeValue;
+    let empty = CqlValue::Empty;
+    let mut v = Vec::new();
+    SerializeValue::serialize(
+        &empty,
+        &ColumnType::Native(NativeType::Ascii),
+        CellWriter::new(&mut v),
+    )
+    .unwrap();
+
+    assert_eq!(v, vec![0, 0, 0, 0]);
+}
+
+#[test]
+fn test_duration_deserialize() {
+    let bytes = [0xc, 0x12, 0xe2, 0x8c, 0x39, 0xd2];
+    let cql_value: CqlValue =
+        super::deser_cql_value(&ColumnType::Native(Duration), &mut &bytes[..]).unwrap();
+    assert_eq!(
+        cql_value,
+        CqlValue::Duration(CqlDuration {
+            months: 6,
+            days: 9,
+            nanoseconds: 21372137
+        })
+    );
+}
+
+#[test]
+fn test_deserialize_empty_payload() {
+    for (test_type, res_cql) in [
+        (ColumnType::Native(Ascii), CqlValue::Ascii("".to_owned())),
+        (ColumnType::Native(Boolean), CqlValue::Empty),
+        (ColumnType::Native(Blob), CqlValue::Blob(vec![])),
+        (ColumnType::Native(NativeType::Counter), CqlValue::Empty),
+        (ColumnType::Native(Date), CqlValue::Empty),
+        (ColumnType::Native(Decimal), CqlValue::Empty),
+        (ColumnType::Native(Double), CqlValue::Empty),
+        (ColumnType::Native(Float), CqlValue::Empty),
+        (ColumnType::Native(Int), CqlValue::Empty),
+        (ColumnType::Native(BigInt), CqlValue::Empty),
+        (ColumnType::Native(Text), CqlValue::Text("".to_owned())),
+        (ColumnType::Native(Timestamp), CqlValue::Empty),
+        (ColumnType::Native(Inet), CqlValue::Empty),
+        (
+            ColumnType::Collection {
+                frozen: false,
+                typ: CollectionType::List(Box::new(ColumnType::Native(Int))),
+            },
+            CqlValue::Empty,
+        ),
+        (
+            ColumnType::Collection {
+                frozen: false,
+                typ: CollectionType::Map(
+                    Box::new(ColumnType::Native(Int)),
+                    Box::new(ColumnType::Native(Int)),
+                ),
+            },
+            CqlValue::Empty,
+        ),
+        (
+            ColumnType::Collection {
+                frozen: false,
+                typ: CollectionType::Set(Box::new(ColumnType::Native(Int))),
+            },
+            CqlValue::Empty,
+        ),
+        (
+            ColumnType::UserDefinedType {
+                frozen: false,
+                definition: Arc::new(UserDefinedType {
+                    name: "".into(),
+                    keyspace: "".into(),
+                    field_types: vec![],
+                }),
+            },
+            CqlValue::Empty,
+        ),
+        (ColumnType::Native(SmallInt), CqlValue::Empty),
+        (ColumnType::Native(TinyInt), CqlValue::Empty),
+        (ColumnType::Native(Time), CqlValue::Empty),
+        (ColumnType::Native(Timeuuid), CqlValue::Empty),
+        (ColumnType::Tuple(vec![]), CqlValue::Empty),
+        (ColumnType::Native(Uuid), CqlValue::Empty),
+        (ColumnType::Native(Varint), CqlValue::Empty),
+    ] {
+        let cql_value: CqlValue = super::deser_cql_value(&test_type, &mut &[][..]).unwrap();
+
+        assert_eq!(cql_value, res_cql);
+    }
+}
+
+#[test]
+fn test_timeuuid_deserialize() {
+    // A few random timeuuids generated manually
+    let tests = [
+        (
+            "8e14e760-7fa8-11eb-bc66-000000000001",
+            [
+                0x8e, 0x14, 0xe7, 0x60, 0x7f, 0xa8, 0x11, 0xeb, 0xbc, 0x66, 0, 0, 0, 0, 0, 0x01,
+            ],
+        ),
+        (
+            "9b349580-7fa8-11eb-bc66-000000000001",
+            [
+                0x9b, 0x34, 0x95, 0x80, 0x7f, 0xa8, 0x11, 0xeb, 0xbc, 0x66, 0, 0, 0, 0, 0, 0x01,
+            ],
+        ),
+        (
+            "5d74bae0-7fa3-11eb-bc66-000000000001",
+            [
+                0x5d, 0x74, 0xba, 0xe0, 0x7f, 0xa3, 0x11, 0xeb, 0xbc, 0x66, 0, 0, 0, 0, 0, 0x01,
+            ],
+        ),
+    ];
+
+    for (uuid_str, uuid_bytes) in &tests {
+        let cql_val: CqlValue =
+            super::deser_cql_value(&ColumnType::Native(Timeuuid), &mut &uuid_bytes[..]).unwrap();
+
+        match cql_val {
+            CqlValue::Timeuuid(uuid) => {
+                assert_eq!(uuid.as_bytes(), uuid_bytes);
+                assert_eq!(CqlTimeuuid::from_str(uuid_str).unwrap(), uuid);
+            }
+            _ => panic!("Timeuuid parsed as wrong CqlValue"),
+        }
+    }
 }
